Add per path callback option
Add constructor option for "finalAnimation". Checks if it is a function,
or throws error. If it exists, save it.
While mapping, check if finalAnimation exists, and save a boolean to
quickly check if each path has played that animation.
While drawing path, check if finalAnimation has not yet been played and
if path has been completely drawn. If so, execute callback, passing the
DOM element for the current path.